Orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator-a2a-poc"
version = "0.1.0"
description = "POC: A2A-style agent workflow (Diagnoser, Fixer, Support) with Streamlit UI"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "POC" }]
license = { text = "MIT" }
dependencies = [
  "streamlit>=1.33",
  "pydantic>=2.6",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "httpx>=0.27",
  "websockets>=12.0",
  "google-generativeai>=0.7.2",
  "python-dotenv>=1.0",
]

[tool.setuptools]
packages = ["app", "app.core", "app.agents", "app.orchestrator", "app.ui"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
